Ensures `getGlobal` returns correct `dust` object.
In Node.js, `var dust`, in the top-level scope does not declare
a global variable. So in situations where `typeof exports` does
not correctly detect Node environments (require.js's build tool
is an example).

This ensures that `global.dust` is the same object as the
`var dust` in the top level scope.