Failing if the template name is empty
if no name is passed to ```dust.compile (<template>, <name>)``` call, it internally registers the template with the name ```null```, which can never be accessed. This would make it difficult to debug if a particular template is not being loaded. So, it would be appropriate to fail, if the name is not given.