Escape template names
Since the compiler was not interpreting template names when emitting
javascript code with those as literals, and instead merely concatenating
string delimiters onto the end, template names ending in a backslash
or containing a newline or double quotes would cause a parser failure
when loading the compiled template.

This additionally escapes the null character, as some browsers will omit
it during parsing, making the template name not make a round-trip from
what was specified to what is cached.