Fix the output of format (whitespace-only) blocks inside inline partials.
When whitespace compression is on, format blocks are optimized away.
When whitespace compression is off, format blocks are usually coerced into buffers.
But if there are no buffers (which is basically only possible within an inline partial, since you're not going to compile a template that's a bunch of newlines and nothing else), we visit `compiler.nodes.format`, which didn't get updated as part of #531 since it's such an edge case.

Closes #556