Refactor main Dust module's global context
Uses the same UMD pattern as parser and compiler now, instead of a slightly different one. `root` should only be used to refer to the browser window, not Node's `global`, so now we use `console` directly instead of `root.console`.

Fixed up tests to ensure that we call dust.log even though debugLevel is turned off to avoid spam, to make sure that we test the code path.

Fixes #500 by removing the IIFE in favor of `this`.
Fixes #524 by allowing DEBUG=dust to be set on the command line to turn on debugging.