Add `chunk.stream` to allow streamables in context
When Dust finds a streamable in context, it begins capturing the streamable's `data` events and renders the block once for each thunk. If the streamable is a reference, Dust writes the thunks directly into the chunk.

Other fixes as part of this commit;
* Fix async tests in the Node runner
* Allow functions in Dust context to return functions recursively

Note: highland.js (browser version) is included as a test lib because the highland repo only exports the Node version as part of its NPM package.